+++ pkg/attestation/crafter/materials/aicodingsession/aicodingsession.go
