+++ pkg/attestation/crafter/materials/chainloop_ai_coding_session.go
@@ -22,9 +22,9 @@ import (
 	"os"
 
 	schemaapi "github.com/chainloop-dev/chainloop/app/controlplane/api/workflowcontract/v1"
-	"github.com/chainloop-dev/chainloop/internal/aicodingsession"
 	"github.com/chainloop-dev/chainloop/internal/schemavalidators"
 	api "github.com/chainloop-dev/chainloop/pkg/attestation/crafter/api/attestation/v1"
+	"github.com/chainloop-dev/chainloop/pkg/attestation/crafter/materials/aicodingsession"
 	"github.com/chainloop-dev/chainloop/pkg/casclient"
 
 	"github.com/rs/zerolog"
@@ -58,8 +58,7 @@ func (c *ChainloopAICodingSessionCrafter) Craft(ctx context.Context, artifactPat
 
 	// Unmarshal envelope, keeping data as raw JSON for schema validation
 	var envelope struct {
-		Schema string          `json:"schema"`
-		Data   json.RawMessage `json:"data"`
+		Data json.RawMessage `json:"data"`
 	}
 	if err := json.Unmarshal(f, &envelope); err != nil {
 		c.logger.Debug().Err(err).Msg("error decoding file")
@@ -87,11 +86,6 @@ func (c *ChainloopAICodingSessionCrafter) Craft(ctx context.Context, artifactPat
 	material, err := uploadAndCraft(ctx, c.input, c.backend, artifactPath, c.logger)
 	if err != nil {
 		return nil, err
-	}
-
-	// Surface schema as an annotation
-	if envelope.Schema != "" {
-		material.Annotations[annotationEvidenceSchema] = envelope.Schema
 	}
 
 	// Surface agent name as an annotation
